# frontend/src/lib/api/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "クラウドストレージシステム GC Storage の REST API",
        "title": "GC Storage API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/auth/email/resend": {
            "post": {
                "description": "メールアドレス確認用のメールを再送信します",
                "tags": [
                    "Auth"
                ],
                "summary": "確認メール再送",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/github_com_Hiro-mackay_gc-storage_backend_internal_interface_dto_request.ResendEmailVerificationRequest"
                            }
                        }
                    },
                    "description": "再送先メールアドレス",
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_interface_handler.SwaggerResendEmailVerificationResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_interface_handler.SwaggerErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/email/verify": {
            "post": {
                "description": "トークンを使用してメールアドレスの確認を行います",
                "tags": [
                    "Auth"
                ],
                "summary": "メールアドレス確認",
                "parameters": [
                    {
                        "description": "メール確認トークン",
                        "name": "token",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_interface_handler.SwaggerVerifyEmailResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_interface_handler.SwaggerErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "description": "メールアドレスとパスワードでログインし、セッションCookieを発行します",
                "tags": [
                    "Auth"
                ],
                "summary": "ログイン",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/github_com_Hiro-mackay_gc-storage_backend_internal_interface_dto_request.LoginRequest"
                            }
                        }
                    },
                    "description": "ログイン情報",
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_interface_handler.SwaggerLoginResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_interface_handler.SwaggerErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_interface_handler.SwaggerErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "security": [
                    {
                        "SessionCookie": []
                    }
                ],
                "description": "現在のセッションを無効化し、セッションCookieを削除します",
                "tags": [
                    "Auth"
                ],
                "summary": "ログアウト",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_interface_handler.SwaggerLogoutResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_interface_handler.SwaggerErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/oauth/{provider}": {
            "post": {
                "description": "OAuth認証コードを使用してログインまたは新規登録します",
                "tags": [
                    "Auth"
                ],
                "summary": "OAuthログイン",
                "parameters": [
                    {
                        "description": "OAuthプロバイダー (google, github)",
                        "name": "provider",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/github_com_Hiro-mackay_gc-storage_backend_internal_interface_dto_request.OAuthLoginRequest"
                            }
                        }
                    },
                    "description": "OAuth認証コード",
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_interface_handler.SwaggerOAuthLoginResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_interface_handler.SwaggerErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_interface_handler.SwaggerErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/password/change": {
            "post": {
                "security": [
                    {
                        "SessionCookie": []
                    }
                ],
                "description": "現在のパスワードを確認した上で新しいパスワードに変更します",
                "tags": [
                    "Auth"
                ],
                "summary": "パスワード変更",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/github_com_Hiro-mackay_gc-storage_backend_internal_interface_dto_request.ChangePasswordRequest"
                            }
                        }
                    },
                    "description": "現在のパスワードと新しいパスワード",
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_interface_handler.SwaggerChangePasswordResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_interface_handler.SwaggerErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_interface_handler.SwaggerErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/password/forgot": {
            "post": {
                "description": "パスワードリセット用のメールを送信します",
                "tags": [
                    "Auth"
                ],
                "summary": "パスワードリセット要求",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/github_com_Hiro-mackay_gc-storage_backend_internal_interface_dto_request.ForgotPasswordRequest"
                            }
                        }
                    },
                    "description": "パスワードリセット対象のメールアドレス",
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_interface_handler.SwaggerForgotPasswordResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_interface_handler.SwaggerErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/password/reset": {
            "post": {
                "description": "トークンを使用してパスワードをリセットします",
                "tags": [
                    "Auth"
                ],
                "summary": "パスワードリセット実行",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/github_com_Hiro-mackay_gc-storage_backend_internal_interface_dto_request.ResetPasswordRequest"
                            }
                        }
                    },
                    "description": "リセットトークンと新しいパスワード",
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_interface_handler.SwaggerResetPasswordResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_interface_handler.SwaggerErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/password/set": {
            "post": {
                "security": [
                    {
                        "SessionCookie": []
                    }
                ],
                "description": "OAuthのみで登録したユーザーがパスワードを新規設定します",
                "tags": [
                    "Auth"
                ],
                "summary": "パスワード設定（OAuth専用ユーザー向け）",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/github_com_Hiro-mackay_gc-storage_backend_internal_interface_dto_request.SetPasswordRequest"
                            }
                        }
                    },
                    "description": "設定するパスワード",
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_interface_handler.SwaggerSetPasswordResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_interface_handler.SwaggerErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_interface_handler.SwaggerErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "メールアドレスとパスワードで新規ユーザーを登録します",
                "tags": [
                    "Auth"
                ],
                "summary": "ユーザー登録",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/github_com_Hiro-mackay_gc-storage_backend_internal_interface_dto_request.RegisterRequest"
                            }
                        }
                    },
                    "description": "登録情報",
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_interface_handler.SwaggerRegisterResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_interface_handler.SwaggerErrorResponse"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_interface_handler.SwaggerErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "servers": [
        {
            "url": "http://localhost:8080/api/v1"
        },
        {
            "url": "https://localhost:8080/api/v1"
        }
    ],
    "components": {
...
        "securitySchemes": {
            "SessionCookie": {
                "type": "apiKey",
                "name": "session_id",
                "in": "cookie"
            }
        },
        "schemas": {
...
            "github_com_Hiro-mackay_gc-storage_backend_internal_interface_dto_request.OAuthLoginRequest": {
                "type": "object",
                "required": [
                    "code"
                ],
                "properties": {
                    "code": {
                        "type": "string"
                    }
                }
            },
...
            "github_com_Hiro-mackay_gc-storage_backend_internal_interface_dto_request.ResendEmailVerificationRequest": {
                "type": "object",
                "required": [
                    "email"
                ],
                "properties": {
                    "email": {
                        "type": "string"
                    }
                }
            },
...
            "internal_interface_handler.SwaggerErrorResponse": {
                "type": "object",
                "properties": {
                    "error": {
                        "$ref": "#/components/schemas/internal_interface_handler.SwaggerErrorDetail"
                    }
                }
            },
...
            "internal_interface_handler.SwaggerResendEmailVerificationResponse": {
                "type": "object",
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/github_com_Hiro-mackay_gc-storage_backend_internal_interface_dto_response.ResendEmailVerificationResponse"
                    },
                    "meta": {
                        "$ref": "#/components/schemas/github_com_Hiro-mackay_gc-storage_backend_internal_interface_presenter.Meta"
                    }
                }
            },
...
            "internal_interface_handler.SwaggerVerifyEmailResponse": {
                "type": "object",
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/github_com_Hiro-mackay_gc-storage_backend_internal_interface_dto_response.VerifyEmailResponse"
                    },
                    "meta": {
                        "$ref": "#/components/schemas/github_com_Hiro-mackay_gc-storage_backend_internal_interface_presenter.Meta"
                    }
                }
            }
        }
    }
}